+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(visvalingam)
+
+test_check("visvalingam")