+++ NEWS.md
@@ -1,3 +1,10 @@
+# visvalingam 0.1.1
+
+* Fix off-by-one error in C code
+* Tidy priority queue implementation
+* Handle duplicates in `vis_simplify()`
+* Duplicate points are given *negative* effective areas in `vis_effective_areas()`
+
 # visvalingam 0.1.0
 
 * Initial release